integrated interp adaptBins

+++ code/kdTree/parallelApproved/adaptBins_new.cpp
@@ -31,13 +31,12 @@ using namespace std;
 //
 
 #include "definitions.h"
-#include "Data.h"
 
 //
 // Function: importFiles
 //
 
-void adaptBins( double *binE, int *binC, int numWorkers, double numLines, int avgPtsPerWorker ) {
+void adaptBins( double *binE, int *binC, int numWorkers, int numLines, double avgPtsPerWorker ) {
 	// binE = bin edges, binC = bin counts
 	
 	double *binE2   = new double[numWorkers+1];    // the new bin edges, copy to binE
@@ -56,10 +55,10 @@ void adaptBins( double *binE, int *binC, int numWorkers, double numLines, int av
 	for( int j = 0; j < numWorkers-1; j++ ) {
 		// get ind
 		ind = 0;
+
 		while( binCum[ind+1] < (j+1)*avgPtsPerWorker ) {
 			ind++;
 		}
-		
 		// get binE2
 		binE2[j+1] = binE[ind] + ( (j+1)*avgPtsPerWorker - binCum[ind] )/binC[ind]*( binE[ind+1] - binE[ind] );
 	}
